RoBERTa-base for News Classification
RoBERTa for News Classification

Here we have used RoBERTa-base Sequence Classifier (Transformer) to classify given news dataset into 4 classes:

○ World (0)
○ Sports (1)
○ Business (2)
○ Sci/Tech (3)

We cleaned the dataset first:

-> Converted the whole dataset to lower case.

-> Removed punctuation marks.

-> Replaced numbers and words containing numbers as a sunstring with "NUM" tokens.


We finetuned this transformer model for 15 epochs to learn the pattern to the best using AdamW optimizer using a LR of 0.00001. We kept a moderate batch size of 16 for frequenet gradient update

The best accuracy score we could achieve on test set is 88.09 % approx.