Neural Network vs. Bi-directional RNN vs Bi-Directional LSTM
News Classification Analysis: Simple Neural Network vs. Bi-directional RNN vs Bi_Directional LSTM

Here we use Simple Neural Network, Bi-directional RNN and Bi_Directional LSTM to classify given news dataset into 4 classes:

○ World (0)
○ Sports (1)
○ Business (2)
○ Sci/Tech (3)

We cleaned the dataset first:

-> Converted the whole dataset to lower case.

-> Removed stop words.

-> Removed punctuation marks.

-> Replaced numbers and words containing numbers as a sunstring with "NUM" tokens.

We have used here word2vec model to convert our news sentences to embeddings instead of the nn.Embedding Layer to learn the same.

In Simple Neural Network, we have used word2vec word embeddings of the words to feed into the network. Using this network we could achieve 82.34 % accuracy on the test set. We trained the model for 25 epochs to learn the pattern to the best using AdamW optimizer using a LR of 0.003

In Bi-RNN we have used nn.Embedding layer but initialized it with the word2vec embeddings and freezed the embedding layer so that it dont receives the gradient updates and learn new embeddings. Here we could achieve an accuracy of 83.25 % on the test set. We trained the model for 20 epochs to learn the pattern to the best using AdamW optimizer using a LR of 0.0001 

In Bi-LSTM we have used nn.Embedding layer but initialized it with the word2vec embeddings and freezed the embedding layer so that it dont receives the gradient updates and learn new embeddings. Here we could achieve an accuracy of 82.23 % on the test set. We trained the model for 20 epochs to learn the pattern to the best using AdamW optimizer using a LR of 0.00005